+++ routes/users.js
@@ -2,19 +2,29 @@ const Router = require('koa-router')
 const bodyParser = require('koa-bodyparser')
 const model = require('../models/usersModel')
 const auth = require('../controllers/auth')
+const can = require('../permissions/usersPermission')
 
 const prefix = '/api/v1/users'
 const router = Router({ prefix: prefix });
 
 // /
 //router.get('/', auth, getAll)
-router.get('/', getAllUsers)
+router.get('/', auth, getAllUsers)
 router.post('/', bodyParser(), createUser)
 // /id
-router.get('/:id([0-9]{1,})', getUserById)
-router.del('/:id([0-9]{1,})', deleteUserById)
+router.get('/:id([0-9]{1,})', auth, getUserById)
+router.del('/:id([0-9]{1,})', auth, deleteUserById)
 // /login
-router.post('/login', loginUser)
+router.post('/login', auth, loginUser)
+
+async function loginUser(ctx) {
+  // return any details needed by the client
+  const { id, username, role } = ctx.state.user
+  const links = {
+    self: `https://${ctx.host}${prefix}/${id}`
+  }
+  ctx.body = { id, username, role, links };
+}
 
 async function createUser(ctx) {
   const body = ctx.request.body
@@ -29,41 +39,44 @@ async function createUser(ctx) {
 }
 
 async function getAllUsers(ctx, next) {
-  // const permission = can.readAll(ctx.state.user);
-  // if (!permission.granted) {
-  //   ctx.status = 403;
-  // } else {
-    //let users = await model.getAllUsers(20, 1)
+  const permission = can.readAll(ctx.state.user);
+  if (!permission.granted) {
+    ctx.status = 403;
+  } else {
     let users = await model.getAllUsers()
     if (users.length) {
       ctx.body = users
     }
-  //}
+  }
 }
 
 async function getUserById(ctx) {
-  let id = ctx.params.id
-  let user = await model.getUserById(id)
-  if (user.length) {
-    ctx.body = user[0]
+  const permission = can.read(ctx.state.user, ctx.params)
+  if (!permission.granted) {
+    ctx.status = 403;
+  }
+  else {
+    let id = ctx.params.id
+    let user = await model.getUserById(id)
+    if (user.length) {
+      ctx.body = user[0]
+    }
   }
 }
 
-async function loginUser(ctx) {
-  // return any details needed by the client
-  const {id, username, role} = ctx.state.user
-  const links = {
-    self: `https://${ctx.host}${prefix}/${id}`
+async function deleteUserById(ctx) {
+  const permission = can.delete(ctx.state.user, ctx.params)
+  if (!permission.granted) {
+    ctx.status = 403;
   }
-  ctx.body = {id, username, role, links};
+  else {
+    let id = ctx.params.id
+    let result = await model.deleteUserById(id)
+    if (result) {
+      ctx.status = 201
+      ctx.body = `Deleted user with id ${id}`
+    }
+  }
 }
 
-async function deleteUserById(ctx){
-  let id = ctx.params.id
-  let result = await model.deleteUserById(id)
-  if (result){
-    ctx.status = 201
-    ctx.body = `Deleted user with id ${id}`
-  }
-}
 module.exports = router;
